Renamed crossModule fodler to shared (DH-17600)
