+++ src/dh/dhe.ts
@@ -39,7 +39,7 @@ import {
   UnsupportedFeatureQueryError,
 } from '../common';
 import { withResolvers } from '../util';
-import type { QuerySerial } from '../crossModule';
+import type { QuerySerial } from '../shared';
 
 export type IDraftQuery = EditableQueryInfo & {
   isClientSide: boolean;
