+++ src/services/ParseDocumentCache.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { URIMap } from './URIMap';
-import { assertDefined } from '../crossModule';
+import { assertDefined } from '../shared';
 
 /**
  * Cache for parsed documents. Cached based on document uri + version.
