+++ src/shared/webViewContentUtils/getExternalThemeData.spec.ts
