+++ src/shared/webViewContentUtils/getVscodeProperty.spec.ts
