+++ src/shared/webViewContentUtils/getVscodeProperty.ts
@@ -1,7 +1,7 @@
 /**
  * Code in this module needs to be consumable from both the extension code (CJS)
  * and the webview content code (ESM). Avoid importing anything here from outside
- * of the `crossModule` folder to minimize the risk of breaking the builds.
+ * of the `shared` folder to minimize the risk of breaking the builds.
  */
 
 import type { WebviewApi } from 'vscode-webview';
