+++ src/types/serviceTypes.d.ts
@@ -25,7 +25,7 @@ import type {
   UnauthenticatedClient as DheUnauthenticatedClient,
   Username,
 } from '@deephaven-enterprise/auth-nodejs';
-import type { QuerySerial } from '../crossModule';
+import type { QuerySerial } from '../shared';
 
 export interface IAsyncCacheService<TKey, TValue> extends IDisposable {
   get: (key: TKey) => Promise<TValue>;
